Struct gemini::astronomicals::galaxy::Galaxy

pub struct Galaxy {
    pub sectors: Vec<Sector>,
    pub map: RTree<Point>,
    pub systems: HashMap<Point, System>,
}

A galaxy of systems.

Fields
sectors: Vec<Sector> map: RTree<Point> systems: HashMap<Point, System>
Methods

impl Galaxy

pub fn new(sectors: Vec<Sector>, systems: Vec<System>) -> Self

Create a new galaxy with the given sectors and systems.

pub fn system(&self, location: &Point) -> Option<&System>

Returns a reference to the system at the given location.

pub fn system_mut(&mut self, location: &Point) -> Option<&mut System>

Returns a mutable reference to the system at the given location.

pub fn systems(
&self
) -> impl Iterator<Item = &System>

Returns references to all systems.

pub fn systems_mut(
&mut self
) -> impl Iterator<Item = &mut System>

Returns mutable references to all systems.

pub fn search_name(&self, query: &str) -> Option<&System>

Finds the system with the closest matching name.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn reachable(&self, location: &Point, max_distance: f64) -> Vec<&Point>

Returns all system locations reachable from the given location within the given radius.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn reachable_rect(&self, corner1: &Point, corner2: &Point) -> Vec<&Point>

Returns all system locations reachable from the given location within the given rectangle.

pub fn nearest(&self, location: &Point) -> Option<&Point>

Returns the nearest system location to the given point.

pub fn route(
&self,
start: Point,
goal: Point,
range: f64,
max_steps: u32
) -> Option<(u32, Vec<Point>)>

Finds the shortest path from start to goal with at most range along any edge and a maximum max_steps number of nodes visited.