Module gemini::astronomicals::system

Re-exports

use super::*;

use entities::Faction;

use game::Updatable;

use std::fmt;

use std::hash::Hash;

use std::hash::Hasher;

use utils::Point;

Structs

Reputation
Represents the current player level of reputation with the system.

System
Represets a single star system with at a given location with the given star and planets.

SystemBuilder
Builder for System.

Enums

SystemSecurity
Represents the different security levels a system is in at a given point.

SystemState
Represents the different states a system is in at a given point.

Constants

_IMPL_DESERIALIZE_FOR_Reputation

_IMPL_DESERIALIZE_FOR_System

_IMPL_DESERIALIZE_FOR_SystemSecurity

_IMPL_DESERIALIZE_FOR_SystemState

_IMPL_SERIALIZE_FOR_Reputation

_IMPL_SERIALIZE_FOR_System

_IMPL_SERIALIZE_FOR_SystemSecurity

_IMPL_SERIALIZE_FOR_SystemState