Struct gemini::economy::agent::Agent

pub struct Agent {
    seed: u32,
    credits: u64,
    populations: Vec<f64>,
    inventory: HashMap<Commodity, u64>,
    ideals: Vec<HashMap<Commodity, u64>>,
    productions: Vec<HashMap<Commodity, u64>>,
    price_beliefs: HashMap<Commodity, Range<i64>>,
}

Economic agent, able to take part in transactions.

Fields
seed: u32 credits: u64 populations: Vec<f64> inventory: HashMap<Commodity, u64> ideals: Vec<HashMap<Commodity, u64>> productions: Vec<HashMap<Commodity, u64>> price_beliefs: HashMap<Commodity, Range<i64>>
Methods

impl Agent

const STARTING_CREDITS: u64

STARTING_CREDITS: u64 = 100000

const INITIAL_LOWER_BELIEF: i64

INITIAL_LOWER_BELIEF: i64 = 1000

const INITIAL_UPPER_BELIEF: i64

INITIAL_UPPER_BELIEF: i64 = 10000

const DEFAULT_PRICE: i64

DEFAULT_PRICE: i64 = (<Agent>::INITIAL_LOWER_BELIEF + <Agent>::INITIAL_UPPER_BELIEF) / 2

const MIN_PRICE: i64

MIN_PRICE: i64 = 10

const POPULATION_FACTOR: f64

POPULATION_FACTOR: f64 = 1.

pub fn new(system: &System) -> Self

Create a new economic agent for the given system.

fn initial_population(mass: f64, kind: &PlanetType) -> f64

Calculates the initial planet population based on mass and planet type.

pub fn hash(&self) -> u32

Returns the hash of the system to which the agent is associated.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn populations(&self) -> Vec<f64>

Returns the system populations.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn prices(&self) -> Vec<(Commodity, i64)>

Returns the prices for all commodities known.

fn balance(&self, commodity: &Commodity) -> i64

Returns the balance of a given commodity compared to the ideal amount currently in inventory.

pub fn update_inventory(&mut self, commodity: &Commodity, delta: i64)

Adds delta amount of the given commodity to the inventory.

pub fn update_credits(&mut self, delta: i64)

Adds delta amount of credits.

pub fn update_price_belief(
&mut self,
commodity: &Commodity,
unit_price: u64,
success: bool
)

Updates the price beliefs for the given commodity based on the given unit price. Success indicates whether the agent was successful in trading. Successful trades will strengthen the belief whereas unsucessful trades will weaken the belief.

pub fn generate_bid(&self, commodity: &Commodity) -> Option<BidBuilder>

Generate a bid for the given commodity if there exists a demand.

pub fn generate_ask(&self, commodity: &Commodity) -> Option<AskBuilder>

Generate an ask for the given commodity if there exists surplus.

pub fn update_population(&mut self, demand_supply: &[(Commodity, u64, u64)])

Updates the planet populations held by the agent based on potential economic growth.