Struct gemini::economy::market::Market

pub struct Market {
    average_prices: HashMap<Commodity, u64>,
    agents: Vec<Arc<Mutex<Agent>>>,
}

Controls an economic market, i.e a sector of trading systems.

Fields
average_prices: HashMap<Commodity, u64> agents: Vec<Arc<Mutex<Agent>>>
Methods

impl Market

pub fn new() -> Self

Creates a new empty market.

pub fn agent(&self, system_hash: u32) -> Option<&Arc<Mutex<Agent>>>

Returns the agent, if any, which is associated with the given system.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn agents(&self) -> &Vec<Arc<Mutex<Agent>>>

Returns a reference to all agents.

pub fn add_system(&mut self, system: &System)

Adds the given system to this market.

fn resolve_offers(
&mut self,
commodity: &Commodity,
bids: Vec<Bid>,
asks: Vec<Ask>
)

Attemps to resolve the bids and asks for the given commodity by matching the highest asks with the lowest bids performing the transaction. Returns the quantity traded.