Module gemini::event

Re-exports

use game::Game;

use std::io;

use std::sync::mpsc::channel;

use std::sync::mpsc::Receiver;

use std::sync::mpsc::Sender;

use std::sync::Arc;

use std::sync::Mutex;

use std::thread::park_timeout;

use std::thread::spawn;

use std::time::Duration;

use std::time::Instant;

use termion::event;

use termion::input::TermRead;

Structs

EventHandler
Accepts events from all senders and propagates them to all recievers.

HANDLER

Enums

Event
User and system events.

Statics

HANDLER

Functions

add_keyboard_handler
Start listener for keyboard events and forward to event handler.

add_player_handler
Start listener for events that should update player state.

add_update_handler
Start listener for events that should run an update on the game state.