Struct gemini::game::Game

pub struct Game {
    pub galaxy: Mutex<Galaxy>,
    pub shipyard: Mutex<Shipyard>,
    pub player: Mutex<Player>,
    pub economy: Mutex<Economy>,
    updated: Mutex<DateTime<Utc>>,
}

Main game state object, shared and syncronized by use of Arc and Mutex.

Fields
galaxy: Mutex<Galaxy> shipyard: Mutex<Shipyard> player: Mutex<Player> economy: Mutex<Economy> updated: Mutex<DateTime<Utc>>
Methods

impl Game

pub fn new() -> Arc<Self>

Creates a new game.

pub fn update(&self)

Update Game information, may advance time.

fn attempt_advance_time(&self) -> Option<i64>

Attemps to advance time returning the number of days advanced if any.

pub fn save_all(&self)

Creates and stores a quicksave of the current game.

pub fn save_player(&self)

Creates and stores a quicksave of the player data.

pub fn load() -> Option<Arc<Self>>

Attempts to load a quicksave of a game state.