Struct gemini::game_config::GameConfig

pub struct GameConfig {
    pub map_seed: u32,
    pub starting_credits: u32,
    pub number_of_systems: u64,
    pub system_spread: f64,
    pub number_of_sectors: usize,
}

Contains high level configuration parameters for the game such as constants for generation.

Fields
map_seed: u32 starting_credits: u32 number_of_systems: u64 system_spread: f64 number_of_sectors: usize
Methods

impl GameConfig

pub fn retrieve() -> GameConfig

Attempts to load a GameConfig from disk at the default preference location. If the loading fails for any reason, for example, the file does not exist, a new default GameConfig object is created, stored and returned.

pub fn store(&self)

Attempts to store the GameConfig on disk at the default preference location.