Struct gemini::generators::planets::PlanetGen

pub struct PlanetGen {
    mass_gen: Exponential,
    orbit_dist_gen: Gamma,
}

Basic non deterministic name generator for generating new Planets which are similar to the trained data provided.

Fields
mass_gen: Exponential orbit_dist_gen: Gamma
Methods

impl PlanetGen

const MIN_DIST: f64

MIN_DIST: f64 = 500.

Minimal distance to star.

pub fn new() -> Self

Create a new Planet generator which loads the star resources needed.

pub fn calculate_surface_temperature(orbit_distance: f64, star: &Star) -> f64

Calculate planet surface temperature from star luminosity and distance to it. Uses the Bond albedo for the Earth.

pub fn predict_type<R: Rng>(
rng: &mut R,
surface_temperature: f64,
mass: f64
) -> PlanetType

Predict the planet type based on surface_temperature and mass.

pub fn predict_economy<R: Rng>(rng: &mut R, kind: &PlanetType) -> PlanetEconomy

Predict the planet economy type based on type.

pub fn generate<R: Rng>(&self, gen: &mut R) -> Option<PlanetBuilder>

Generates a new PlanetBuilder from the distribution using the provided random generator. Sets the fields which are independent on the context.