Trait gemini::gui::tab::Tab

pub trait Tab {
    fn new(_: Arc<Game>, _: Sender<Event>) -> Box<Self>
    where
        Self: Sized;

    fn title(&self) -> String;

    fn handle_event(&mut self, event: Event) -> Option<GUIEvent>;

    fn draw(&self, term: &mut Terminal<MouseBackend>, area: &Rect);
}

Interface for dealing with visual tabs in the GUI.

Required Methods

Important traits for Box<R>
impl<R> Read for Box<R> where
R: Read + ?Sized, impl<W> Write for Box<W> where
W: Write + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

fn new(_: Arc<Game>, _: Sender<Event>) -> Box<Self> where
Self: Sized,

Creates a new tab.

fn title(&self) -> String

Returns the title string describing the tab.

fn handle_event(&mut self, event: Event) -> Option<GUIEvent>

Handles the user provided event.

fn draw(&self, term: &mut Terminal<MouseBackend>, area: &Rect)

Draws the tab in the given terminal and area.