Struct gemini::player::Player

pub struct Player {
    credits: u32,
    ship: Option<Ship>,
    location: Point,
    state: PlayerState,
}

Player type holding the player's current ship, credits etc.

Fields
credits: u32 ship: Option<Ship> location: Point state: PlayerState
Methods

impl Player

const TRAVEL_SPEED: f64

TRAVEL_SPEED: f64 = 10. / 60000.

Travling speed between systems, ly/ms.

pub fn new(credits: u32, ship: Ship, location: &Point) -> Self

Create a new player.

pub fn update_state(&mut self)

Update the player state.

pub fn balance(&self) -> u32

Returns the player's current balance.

pub fn ship(&self) -> &Option<Ship>

Returns an reference to the player's active ship.

pub fn ship_mut(&mut self) -> &mut Option<Ship>

Returns an reference to the player's active ship.

pub fn location(&self) -> Point

Get the current player location.

pub fn state(&self) -> PlayerState

Returns the player state.

pub fn dock(&mut self, planet_id: usize)

Docks the player to the planet with the given id.

pub fn undock(&mut self)

Undocks the player from its current planet.

pub fn refuel(&mut self)

Attemps to fuel up the player ship as far as credits reaches.

pub fn set_route(&mut self, route: Vec<Point>)

Sets the route for the player.

pub fn route(&self) -> Option<Vec<&Point>>

Get the player's currrent route, if available.

pub fn eta(&self) -> Option<(String, Point)>

Get the estimated time of arrival (local time) and the destination of the current route.