Module gemini::resources

Re-exports

use serde::de::Deserialize;

use serde_json;

use std::collections::HashMap;

use std::str;

use astronomicals::planet::PlanetEconomy;

use economy::Commodity;

use entities::Faction;

use ship::ShipCharacteristics;

Structs

AgentResource
Resource containing all production/consumptions for factions and planets.

AstronomicalNamesResource
Resource used in name generation of celestial objects.

RESOURCES

ShipResource
Resource with all ships available in the game.

Constants

_IMPL_DESERIALIZE_FOR_AgentResource

_IMPL_DESERIALIZE_FOR_AstronomicalNamesResource

_IMPL_DESERIALIZE_FOR_ShipResource

_IMPL_SERIALIZE_FOR_AgentResource

_IMPL_SERIALIZE_FOR_AstronomicalNamesResource

_IMPL_SERIALIZE_FOR_ShipResource

Statics

RESOURCES

Traits

Resource
Generic Resource trait to be implemented by all resource types which should be loaded at compile time. KEY must be unique to the specific resource (e.g the filename of the resource).

Functions

fetch_resource
Attempts to returns the resource with the given type, will return None if the type has no resource or if the deserialization fails.