Struct gemini::utils::Point

pub struct Point {
    pub x: f64,
    pub y: f64,
}

Generic Point type for geometry.

Fields
x: f64 y: f64
Methods

impl Point

pub fn new(x: f64, y: f64) -> Point

Create a new point.

pub fn origin() -> Point

Create a new point with origin coordinates.

pub fn distance(&self, other: &Point) -> f64

Returns the euclidian distance to another point.

pub fn hash(&self) -> u64

Returns the hash of the point coordinates. Hash based on algorithm used is presented in the paper: Optimized Spatial Hashing for Collision Detection of Deformable Objects.